♻️  refactor: rename pages for consistent naming (#11)
